Modularize video section
